requirements.txt
torch
torchvision
torchaudio
Pillow
numpy
tqdm
opencv-python
